+++ examples/cluster/gems.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+# Released under the MIT License.
+# Copyright, 2026, by Samuel Williams.
+
+source "https://rubygems.org"
+
+gem "falcon", "~> 0.56.0"
+gem "async-service-supervisor-envoy", "~> 0.2.0"